Move the AHP and GlobalWeights components
