+++ frontend/src/components/ahp/AhpComparison.tsx
