+++ frontend/src/components/ahp/AhpResultsTable.tsx
