+++ frontend/src/components/globalWeights/GlobalWeightRunRow.tsx
