+++ frontend/src/components/globalWeights/GlobalWeightRunTable.tsx
