+++ frontend/src/components/globalWeights/GlobalWeightsHeader.tsx
